fixed database so now can handle floats

+++ app.py
@@ -8,8 +8,8 @@ db = SQLAlchemy(app)
 
 class Data(db.Model):
 	id = db.Column(db.Integer, primary_key = True)
-	xValues = db.Column(db.Integer)
-	yValues = db.Column(db.Integer)
+	xValues = db.Column(db.Float)
+	yValues = db.Column(db.Float)
 	valid = db.Column(db.Integer, default = 1)
 	date_created = db.Column(db.DateTime, default = datetime.utcnow)
 
@@ -21,8 +21,8 @@ class Data(db.Model):
 def index():
 	if request.method == 'POST':
 		task_content =request.form['content']
-		xVal = int(task_content.split(',')[0])
-		yVal = int(task_content.split(',')[1])
+		xVal = float(task_content.split(',')[0])
+		yVal = float(task_content.split(',')[1])
 		new_point = Data(xValues = xVal, yValues = yVal)
 
 		try:
@@ -61,8 +61,8 @@ def graph():
 		print('y points')
 		for y in yPoints:
 			print(y)
-		# print(xPoints)
-		# print(yPoints)
+		# # print(xPoints)
+		# # print(yPoints)
 		return "worked"
 	else:
 		return render_template('index.html', points = points)
